+++ src/lib/drivers/accelerometer/PX4Accelerometer.hpp
@@ -79,7 +79,7 @@ private:
 	float			_scale{1.f};
 	float			_temperature{NAN};
 
-	float			_clip_limit{_range / _scale};
+	int16_t			_clip_limit{(int16_t)(_range / _scale)};
 
 	uint32_t		_error_count{0};
