+++ src/modules/navigator/mission.cpp
@@ -1892,8 +1892,8 @@ void Mission::publish_navigator_mission_item()
 	navigator_mission_item.instance_count = _navigator->mission_instance_count();
 	navigator_mission_item.sequence_current = _current_mission_index;
 	navigator_mission_item.nav_cmd = _mission_item.nav_cmd;
-	navigator_mission_item.latitude = _mission_item.lat;
-	navigator_mission_item.longitude = _mission_item.lon;
+	navigator_mission_item.latitude = (float)_mission_item.lat;
+	navigator_mission_item.longitude = (float)_mission_item.lon;
 	navigator_mission_item.altitude = _mission_item.altitude;
 
 	navigator_mission_item.time_inside = get_time_inside(_mission_item);
